# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "EmployeeDTO" : {
        "description" : "An employee",
        "properties" : {
          "address" : {
            "description" : "Only for managers and the employee themselves, otherwise null",
            "example" : "Mainz, Germany",
            "nullable" : true,
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "department" : {
            "example" : "IT",
            "type" : "string"
          },
          "email" : {
            "example" : "jose@example.com",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "description" : "Display name",
            "example" : "José Antonio Cid",
            "type" : "string"
          },
          "role" : {
            "description" : "Job title",
            "example" : "Java Senior Backend",
            "type" : "string"
          },
          "salary" : {
            "description" : "Only for managers and the employee themselves, otherwise null",
            "example" : 75600.0,
            "format" : "double",
            "nullable" : true,
            "type" : "number"
          },
          "username" : {
            "description" : "Keycloak username (lower case), never changes",
            "example" : "jose",
            "type" : "string"
          }
        },
        "required" : [ "address", "createdAt", "department", "email", "id", "name", "role", "salary", "username" ],
        "type" : "object"
      },
      "EmployeeRequestDTO" : {
        "description" : "Employee to create (every field required) or update (see the operation for the rules)",
        "properties" : {
          "address" : {
            "example" : "Mainz, Germany",
            "type" : "string"
          },
          "department" : {
            "example" : "IT",
            "type" : "string"
          },
          "email" : {
            "example" : "jose@example.com",
            "type" : "string"
          },
          "name" : {
            "example" : "José Antonio Cid",
            "type" : "string"
          },
          "role" : {
            "description" : "Job title",
            "example" : "Java Senior Backend",
            "type" : "string"
          },
          "salary" : {
            "example" : 75600.0,
            "format" : "double",
            "type" : "number"
          },
          "username" : {
            "description" : "Keycloak username: letters, digits, `. _ @ -`. Required on create, immutable afterwards",
            "example" : "jose",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ErrorResponse" : {
        "description" : "Error returned by the API",
        "properties" : {
          "code" : {
            "description" : "HTTP status name",
            "example" : "BAD_REQUEST",
            "type" : "string"
          },
          "message" : {
            "description" : "Human-readable explanation",
            "example" : "message must contain 1 to 500 characters",
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token issued by Keycloak (realm hr-realm, client hr-api-login)",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Peer recognition and HR insights. Employees send (optionally anonymous) feedback to colleagues; managers see aggregated insights and control the AI sentiment analysis.\n\nEvery endpoint needs a Keycloak access token (`Authorization: Bearer <token>`) of the `hr-realm` realm. Roles come from the realm roles `MANAGER` and `EMPLOYEE`. Errors have the body `{\"code\": \"...\", \"message\": \"...\"}`, except 401 and role-based 403 responses, which are produced by Spring Security without a body.",
    "license" : {
      "name" : "MIT",
      "url" : "https://github.com/jacidProgrammer/HRWebApp/blob/main/LICENSE"
    },
    "title" : "HRWebApp API",
    "version" : "2.0.0"
  },
  "openapi" : "3.0.1",
  "paths" : {
    "/employees" : {
      "get" : {
        "description" : "All employees. `salary` and `address` are null unless the caller is a manager or it is their own record.",
        "operationId" : "getEmployees",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EmployeeDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "description" : "Missing, invalid or expired access token (no body)"
          },
          "403" : {
            "description" : "The caller's role is not allowed to use this endpoint (no body)"
          }
        },
        "summary" : "List employees",
        "tags" : [ "Employees" ]
      },
      "post" : {
        "description" : "Every field is required. `username` must match a Keycloak username and is stored in lower case.",
        "operationId" : "createEmployee",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmployeeRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EmployeeDTO"
                }
              }
            },
            "description" : "Created; the Location header points to the new employee"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing or invalid fields"
          },
          "401" : {
            "description" : "Missing, invalid or expired access token (no body)"
          },
          "403" : {
            "description" : "The caller's role is not allowed to use this endpoint (no body)"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An employee with this username already exists"
          }
        },
        "summary" : "Create an employee",
        "tags" : [ "Employees" ]
      }
    },
    "/employees/me" : {
      "get" : {
        "description" : "The employee linked to the token's `preferred_username`.",
        "operationId" : "getCurrentEmployee",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EmployeeDTO"
                }
              }
            },
            "description" : "The caller's employee record"
          },
          "401" : {
            "description" : "Missing, invalid or expired access token (no body)"
          },
          "403" : {
            "description" : "The caller's role is not allowed to use this endpoint (no body)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The caller has no employee record (e.g. the manager demo user)"
          }
        },
        "summary" : "Get the caller's employee record",
        "tags" : [ "Employees" ]
      }
    },
    "/employees/{id}" : {
      "delete" : {
        "description" : "Deletes the feedback about the employee and keeps the feedback they wrote, without author.",
        "operationId" : "deleteEmployee",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "401" : {
            "description" : "Missing, invalid or expired access token (no body)"
          },
          "403" : {
            "description" : "The caller's role is not allowed to use this endpoint (no body)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Employee not found"
          }
        },
        "summary" : "Delete an employee",
        "tags" : [ "Employees" ]
      },
      "get" : {
        "description" : "Same field visibility as the list.",
        "operationId" : "getEmployee",
        "parameters" : [ {
          "description" : "Employee id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EmployeeDTO"
                }
              }
            },
            "description" : "The employee"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The id is not a UUID"
          },
          "401" : {
            "description" : "Missing, invalid or expired access token (no body)"
          },
          "403" : {
            "description" : "The caller's role is not allowed to use this endpoint (no body)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Employee not found"
          }
        },
        "summary" : "Get an employee",
        "tags" : [ "Employees" ]
      },
      "put" : {
        "description" : "Managers replace every field except `username` (which can never change). Employees can only update their own `email` and `address`; omitted fields are kept, other fields must be omitted or unchanged.",
        "operationId" : "updateEmployee",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmployeeRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EmployeeDTO"
                }
              }
            },
            "description" : "The updated employee"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing or invalid fields, or an attempt to change the username"
          },
          "401" : {
            "description" : "Missing, invalid or expired access token (no body)"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The caller's role is not allowed, or an employee updates someone else or a manager-only field"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Employee not found"
          }
        },
        "summary" : "Update an employee",
        "tags" : [ "Employees" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Local development",
    "url" : "http://localhost:8080"
  } ],
...
}